requirements.txt
PyYAML
elasticsearch
elasticsearch_dsl
